+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 -   Percent-encode plus (+) when building URLs and in test requests. :issue:`2657`
 -   Cookie values don't quote characters defined in RFC 6265. :issue:`2659`
 -   Include ``pyi`` files for ``datastructures`` type annotations. :issue:`2660`
+-   ``Authorization`` and ``WWWAuthenticate`` objects can be compared for equality.
+    :issue:`2665`
 
 
 Version 2.3.0